разделил по ЕТ заполнение массива и функцию самого задания
