+++ sobolevsky.misha/P4/fillMatrix.cpp
@@ -1,0 +1,17 @@
+#include "funcs.hpp"
+#include <fstream>
+
+namespace sobolevsky
+{
+  size_t fillMatrix(std::istream & in, int * matrix, size_t s, size_t counter)
+  {
+    for (size_t i = 0; i < s; ++i)
+    {
+      if (in >> matrix[i])
+      {
+        counter++;
+      }
+    }
+    return counter;
+  }
+}
