+++ sobolevsky.misha/P4/funcs.hpp
@@ -4,7 +4,6 @@
 
 namespace sobolevsky
 {
-  size_t fillMatrix(std::istream & in, int * matrix, size_t s, size_t counter);
   void isTrianglMatrix(std::ofstream & out, int * matrix, size_t rows, size_t cols);
 }
